requirements-api.txt
# API dependencies for ebook-tts FastAPI server
#
# Install with: pip install -r requirements-api.txt
# (After installing the main package with: pip install .)

# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
email-validator>=2.0  # Required for Pydantic EmailStr

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1  # Pin version for passlib compatibility

# Database
sqlalchemy>=2.0
psycopg2-binary>=2.9.9  # PostgreSQL driver
# alembic>=1.13.0  # Uncomment if you need migrations

# S3-compatible storage
boto3>=1.34.0

# Settings management
pydantic-settings>=2.1.0
